+++ app/constants.ts
@@ -1,0 +1,10 @@
+export const BG = "#0E0E10"; // deeper, cleaner dark
+export const DOT = "rgba(255,255,255,0.05)"; // subtle dot
+export const VIGNETTE = "radial-gradient(1000px 500px at 50% -10%, rgba(255,255,255,0.03), transparent)";
+export const ACCENT = "#7B61FF"; // calm ultraviolet
+export const ACCENT_HOVER = "#6B54E3";
+export const RING = "rgba(123, 97, 255, 0.45)";
+export const GLASS_BG = "linear-gradient(180deg, rgba(255,255,255,0.06), rgba(255,255,255,0.02))";
+export const BORDER_SOFT = "rgba(255,255,255,0.10)";
+export const TEXT_SOFT = "#CFCFD4";
+export const PREVIEW_IMG = "https://l7y6qjyp5m.ufs.sh/f/usun6XPoM0UCnar5XbcjR2aezOZ4lNvPKq05MfxnY3hisyg1";
